+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ContributtionCard from './components/organisms/ContributtionCard';
 import './App.css';
-import { getContributions } from './functions/GetGithubContribute';
+import { getContributions } from './functions/GetGithubContribution';
 
 const App: React.FC = () => {
   return (
